Full type annotations + linting

+++ gpxtrackposter/track_loader.py
@@ -1,17 +1,19 @@
 """Handle parsing of GPX files and writing/loading of cached data"""
 
 
-# Copyright 2016-2019 Florian Pigorsch & Contributors. All rights reserved.
+# Copyright 2016-2020 Florian Pigorsch & Contributors. All rights reserved.
 #
 # Use of this source code is governed by a MIT-style
 # license that can be found in the LICENSE file.
 
+import concurrent.futures
 import hashlib
 import logging
 import os
 import shutil
-from typing import Dict, Generator, List
-import concurrent.futures
+import typing
+
+
 from gpxtrackposter.exceptions import ParameterError, TrackLoadError
 from gpxtrackposter.track import Track
 from gpxtrackposter.year_range import YearRange
@@ -21,7 +23,7 @@ log = logging.getLogger(__name__)
 
 def load_gpx_file(file_name: str) -> Track:
     """Load an individual GPX file as a track by using Track.load_gpx()"""
-    log.info(f"Loading track {os.path.basename(file_name)}...")
+    log.info("Loading track %s...", os.path.basename(file_name))
     t = Track()
     t.load_gpx(file_name)
     return t
@@ -33,7 +35,7 @@ def load_cached_track_file(cache_file_name: str, file_name: str) -> Track:
         t = Track()
         t.load_cache(cache_file_name)
         t.file_names = [os.path.basename(file_name)]
-        log.info(f"Loaded track {file_name} from cache file {cache_file_name}")
+        log.info("Loaded track %s from cache file %s", file_name, cache_file_name)
         return t
     except Exception as e:
         raise TrackLoadError("Failed to load track from cache.") from e
@@ -53,46 +55,47 @@ class TrackLoader:
         load_tracks: Load all data from cache and GPX files
     """
 
-    def __init__(self):
+    def __init__(self) -> None:
         self.min_length = 1000
-        self.special_file_names = []
+        self.special_file_names: typing.List[str] = []
         self.year_range = YearRange()
-        self.cache_dir = None
-        self._cache_file_names = {}
+        self.cache_dir: typing.Optional[str] = None
+        self._cache_file_names: typing.Dict[str, str] = {}
 
-    def clear_cache(self):
+    def set_cache_dir(self, cache_dir: str) -> None:
+        self.cache_dir = cache_dir
+
+    def clear_cache(self) -> None:
         """Remove cache directory, if it exists"""
-        if os.path.isdir(self.cache_dir):
-            log.info(f"Removing cache dir: {self.cache_dir}")
+        if self.cache_dir is not None and os.path.isdir(self.cache_dir):
+            log.info("Removing cache dir: %s", self.cache_dir)
             try:
                 shutil.rmtree(self.cache_dir)
             except OSError as e:
-                log.error(f"Failed: {e}")
+                log.error("Failed: %s", str(e))
 
-    def load_tracks(self, base_dir: str) -> List[Track]:
+    def load_tracks(self, base_dir: str) -> typing.List[Track]:
         """Load tracks base_dir and return as a List of tracks"""
-        file_names = [x for x in self._list_gpx_files(base_dir)]
-        log.info(f"GPX files: {len(file_names)}")
+        file_names = list(self._list_gpx_files(base_dir))
+        log.info("GPX files: %d", len(file_names))
 
-        tracks = []  # type: List[Track]
+        tracks: typing.List[Track] = []
 
         # load track from cache
-        cached_tracks = {}  # type: Dict[str, Track]
+        cached_tracks: typing.Dict[str, Track] = {}
         if self.cache_dir:
-            log.info(f"Trying to load {len(file_names)} track(s) from cache...")
+            log.info("Trying to load %d track(s) from cache...", len(file_names))
             cached_tracks = self._load_tracks_from_cache(file_names)
-            log.info(f"Loaded tracks from cache: {len(cached_tracks)}")
+            log.info("Loaded tracks from cache: %d", len(cached_tracks))
             tracks = list(cached_tracks.values())
 
         # load remaining gpx files
         remaining_file_names = [f for f in file_names if f not in cached_tracks]
         if remaining_file_names:
-            log.info(
-                f"Trying to load {len(remaining_file_names)} track(s) from GPX files; this may take a while..."
-            )
+            log.info("Trying to load %d track(s) from GPX files; this may take a while...", len(remaining_file_names))
             loaded_tracks = self._load_tracks(remaining_file_names)
             tracks.extend(loaded_tracks.values())
-            log.info(f"Conventionally loaded tracks: {len(loaded_tracks)}")
+            log.info("Conventionally loaded tracks: %d", len(loaded_tracks))
             self._store_tracks_to_cache(loaded_tracks)
 
         tracks = self._filter_tracks(tracks)
@@ -102,25 +105,23 @@ class TrackLoader:
         # filter out tracks with length < min_length
         return [t for t in tracks if t.length >= self.min_length]
 
-    def _filter_tracks(self, tracks: List[Track]) -> List[Track]:
+    def _filter_tracks(self, tracks: typing.List[Track]) -> typing.List[Track]:
         filtered_tracks = []
         for t in tracks:
             file_name = t.file_names[0]
             if t.length == 0:
-                log.info(f"{file_name}: skipping empty track")
+                log.info("%s: skipping empty track", file_name)
             elif not t.start_time:
-                log.info(f"{file_name}: skipping track without start time")
+                log.info("%s: skipping track without start time", file_name)
             elif not self.year_range.contains(t.start_time):
-                log.info(
-                    f"{file_name}: skipping track with wrong year {t.start_time.year}"
-                )
+                log.info("%s: skipping track with wrong year %d", file_name, t.start_time.year)
             else:
                 t.special = file_name in self.special_file_names
                 filtered_tracks.append(t)
         return filtered_tracks
 
     @staticmethod
-    def _merge_tracks(tracks: List[Track]) -> List[Track]:
+    def _merge_tracks(tracks: typing.List[Track]) -> typing.List[Track]:
         log.info("Merging tracks...")
         tracks = sorted(tracks, key=lambda t1: t1.start_time)
         merged_tracks = []
@@ -135,29 +136,26 @@ class TrackLoader:
                 else:
                     merged_tracks.append(t)
             last_end_time = t.end_time
-        log.info(f"Merged {len(tracks) - len(merged_tracks)} track(s)")
+        log.info("Merged %d track(s)", len(tracks) - len(merged_tracks))
         return merged_tracks
 
     @staticmethod
-    def _load_tracks(file_names: List[str]) -> Dict[str, Track]:
+    def _load_tracks(file_names: typing.List[str]) -> typing.Dict[str, Track]:
         tracks = {}
         with concurrent.futures.ProcessPoolExecutor() as executor:
-            future_to_file_name = {
-                executor.submit(load_gpx_file, file_name): file_name
-                for file_name in file_names
-            }
+            future_to_file_name = {executor.submit(load_gpx_file, file_name): file_name for file_name in file_names}
         for future in concurrent.futures.as_completed(future_to_file_name):
             file_name = future_to_file_name[future]
             try:
                 t = future.result()
             except TrackLoadError as e:
-                log.error(f"Error while loading {file_name}: {e}")
+                log.error("Error while loading %s: %s", file_name, str(e))
             else:
                 tracks[file_name] = t
 
         return tracks
 
-    def _load_tracks_from_cache(self, file_names: List[str]) -> Dict[str, Track]:
+    def _load_tracks_from_cache(self, file_names: typing.List[str]) -> typing.Dict[str, Track]:
         tracks = {}
         with concurrent.futures.ProcessPoolExecutor() as executor:
             future_to_file_name = {
@@ -179,21 +177,21 @@ class TrackLoader:
                 tracks[file_name] = t
         return tracks
 
-    def _store_tracks_to_cache(self, tracks: Dict[str, Track]):
+    def _store_tracks_to_cache(self, tracks: typing.Dict[str, Track]) -> None:
         if (not tracks) or (not self.cache_dir):
             return
 
-        log.info(f"Storing {len(tracks)} track(s) to cache...")
+        log.info("Storing %d track(s) to cache...", len(tracks))
         for (file_name, t) in tracks.items():
             try:
                 t.store_cache(self._get_cache_file_name(file_name))
             except Exception as e:
-                log.error(f"Failed to store track {file_name} to cache: {e}")
+                log.error("Failed to store track %s to cache: %s", file_name, str(e))
             else:
-                log.info(f"Stored track {file_name} to cache")
+                log.info("Stored track %s to cache", file_name)
 
     @staticmethod
-    def _list_gpx_files(base_dir: str) -> Generator[str, None, None]:
+    def _list_gpx_files(base_dir: str) -> typing.Generator[str, None, None]:
         base_dir = os.path.abspath(base_dir)
         if not os.path.isdir(base_dir):
             raise ParameterError(f"Not a directory: {base_dir}")
